refactor(blog): relocate blog-related pages to /blog directory
