+++ redux-blog/src/routes/index.jsx
@@ -1,10 +1,10 @@
 import App from "../App.jsx";
 import { createBrowserRouter } from "react-router-dom";
 import MainLayout from "../layouts/MainLayout.jsx";
-import Blogs from "../pages/Blogs.jsx";
-import Blog from "../pages/Blog.jsx";
-import CreateBlog from "../pages/CreateBlog.jsx";
-import EditBlog from "../pages/EditBlog.jsx";
+import Blogs from "../pages/blog/Blogs.jsx";
+import Blog from "../pages/blog/Blog.jsx";
+import CreateBlog from "../pages/blog/CreateBlog.jsx";
+import EditBlog from "../pages/blog/EditBlog.jsx";
 
 export const routes = createBrowserRouter([
   {
